changed display to show Total Order Amount
The Total Order Amount was not displayed before because it was always exactly equal to the input order Amount. However, two new additions have changed that - "Maximum Amount" and "Slope". Both of these settings can change the executed total amount of orders to be different from the input order Amount. So, the actual amount to be executed is now displayed.